renamed sensors to modules
