+++ src/app.js
@@ -3,57 +3,17 @@ var board = new five.Board({port: process.argv[2]});
 var mbot = require("./mbotlayout");
 
 var Robot = require("./robot");
-var startButton = require("./robot/startbutton");
-var lineSensor = require("./robot/linesensor");
-var lightSensor = require("./robot/lightsensor");
-var distanceSensor = require("./robot/distancesensor");
-var lights = require("./robot/lights");
+var startButton = require("./modules/startbutton");
+var followLine = require("./modules/followline");
+var automatedLight = require("./modules/automatedlight");
+var finish = require("./modules/finish");
 
-//var stdin = process.openStdin();
-//require("tty").setRawMode(true);
 
 board.on("ready", function(){
   this.robot = new Robot(board);
-  //startbutton.initialize(this.robot);
-  //followline.initialize(this.robot);
 
-  lineSensor.initialize(this.robot);
-  lightSensor.initialize(this.robot);
-  distanceSensor.initialize(this.robot);
+  followLine.initialize(this.robot);
+  automatedLight.initialize(this.robot);
+  //finish.initialize(this.robot);
   startButton.initialize(this.robot);
 });
-
-/*stdin.on("keypress", function(chunk, key) {
-  // process the keypresses
-
-  //if (key) {
-    switch (key.name) {
-      case "up":
-        wheels.left(1);
-        wheels.right(1);
-        //console.log("test forward");
-        break;
-      case "down":
-        wheels.left(-1);
-        wheels.right(-1);
-        //console.log("test backward");
-        break;
-      case "left":
-        wheels.left(-1);
-        wheels.right(1);
-        //console.log("test left");
-        break;
-      case "right":
-        wheels.left(1);
-        wheels.right(-1);
-        //console.log("test right");
-        break;
-      case "space":
-        wheels.left(0);
-        wheels.right(0);
-        //console.log("test stop");
-        break;
-    }
-  //}
-});
-*/
